fix(ui): keep terminal session contract with reconciliation owner
